# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export type SectShp = typeof SectShp[keyof typeof SectShp];


export const SectShp = {
  ic: 'ic',
  sq: 'sq',
  plq: 'plq',
} as const;

export interface Sect {
  id: string;
  name: string;
  pos: number;
  shp: SectShp;
  crt_at: string;
  upd_at: string;
}

export type NewSctShp = typeof NewSctShp[keyof typeof NewSctShp];


export const NewSctShp = {
  ic: 'ic',
  sq: 'sq',
  plq: 'plq',
} as const;

export interface NewSct {
  /** @minLength 1 */
  name: string;
  shp?: NewSctShp;
}

export type UpdSctShp = typeof UpdSctShp[keyof typeof UpdSctShp];


export const UpdSctShp = {
  ic: 'ic',
  sq: 'sq',
  plq: 'plq',
} as const;

export interface UpdSct {
  /** @minLength 1 */
  name?: string;
  shp?: UpdSctShp;
}

export interface BlkSct {
  /** @minItems 1 */
  items: NewSct[];
}

export interface SectPage {
  data: Sect[];
  pg: number;
  sz: number;
  tot: number;
}

export type ItemShp = typeof ItemShp[keyof typeof ItemShp];


export const ItemShp = {
  ic: 'ic',
  sq: 'sq',
  plq: 'plq',
} as const;

export interface Item {
  id: string;
  sect_id: string;
  pos: number;
  name: string;
  dsc?: string | null;
  price: string;
  img?: string | null;
  shp: ItemShp;
  crt_at: string;
  upd_at: string;
}

export type NewItmShp = typeof NewItmShp[keyof typeof NewItmShp];


export const NewItmShp = {
  ic: 'ic',
  sq: 'sq',
  plq: 'plq',
} as const;

export interface NewItm {
  sect_id: string;
  /** @minLength 1 */
  name: string;
  dsc?: string;
  price?: string;
  img?: string;
  shp?: NewItmShp;
}

export type UpdItmShp = typeof UpdItmShp[keyof typeof UpdItmShp];


export const UpdItmShp = {
  ic: 'ic',
  sq: 'sq',
  plq: 'plq',
} as const;

export interface UpdItm {
  /** @minLength 1 */
  name?: string;
  dsc?: string;
  price?: string;
  img?: string;
  shp?: UpdItmShp;
}

export interface BlkItm {
  /** @minItems 1 */
  items: NewItm[];
}

export interface ItemPage {
  data: Item[];
  pg: number;
  sz: number;
  tot: number;
}

export interface OrdReq {
  /** @minItems 1 */
  ids: string[];
}

export interface OrdItm {
  /** @minItems 1 */
  ids: string[];
  sectId?: string;
}

export type ListSectsParams = {
pg?: number;
sz?: number;
};

export type ListItemsParams = {
sectId: string;
pg?: number;
sz?: number;
};

export type QueryItemsBody = {
  sectIds: string[];
  pg?: number;
  sz?: number;
};