# lib/api-client-react/src/generated/schemas/itmQry.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */

export interface ItmQry {
  /** @minItems 1 */
  sectIds: string[];
  pg?: number;
  sz?: number;
}