# lib/api-client-react/src/generated/schemas/newSct.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
import type { NewSctShp } from './newSctShp';

export interface NewSct {
  /** @minLength 1 */
  name: string;
  shp?: NewSctShp;
}