# lib/api-client-react/src/generated/schemas/ordItm.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */

export interface OrdItm {
  /** @minItems 1 */
  ids: string[];
  sectId?: string;
}