# lib/api-client-react/src/generated/schemas/updSct.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
import type { UpdSctShp } from './updSctShp';

export interface UpdSct {
  /** @minLength 1 */
  name?: string;
  shp?: UpdSctShp;
}