# lib/api-zod/src/generated/types/item.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
import type { ItemShp } from './itemShp';

export interface Item {
  id: string;
  sect_id: string;
  pos: number;
  name: string;
  dsc?: string | null;
  price: string;
  img?: string | null;
  shp: ItemShp;
  crt_at: Date;
  upd_at: Date;
}