# lib/api-zod/src/generated/types/newItm.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
import type { NewItmShp } from './newItmShp';

export interface NewItm {
  sect_id: string;
  /** @minLength 1 */
  name: string;
  dsc?: string;
  price?: string;
  img?: string;
  shp?: NewItmShp;
}