# lib/api-zod/src/generated/types/sect.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
import type { SectShp } from './sectShp';

export interface Sect {
  id: string;
  name: string;
  pos: number;
  shp: SectShp;
  crt_at: Date;
  upd_at: Date;
}